hmp: Print if memory section is registered in KVM
This adds a "KVM" marker to the "into mtree -f" to tell the user if
a particular memory section is registered as a KVM memory slot; useful
for debugging purposes.

Since it is memory sections which get registered in KVM (rather than
memory regions), this only prints "KVM" for flatviews.

For example:
 Root memory region: system
  0000000000000000-00000003ffffffff (prio 0, ram): ppc_spapr.ram KVM
  0000200000000020-000020000000003f (prio 1, i/o): virtio-pci

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>