ppc/vof: Check for quiesce
There is no distinct moment when the firmware is no longer present and
only RTAS is available. The closest is the "quiesce" CI call which is
stopping devices so after this point the client does not call CI anymore
except for printing messages which is not a bug but using CI for anything
else is rather suspicious so add a trace point to catch these when in
doubt.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>