spapr/vof: Add block device boot support
This hooks up disks to block backends so vof.bin can read MBR/GPT,
find a bootloader and run it. This bypasses the device drivers
and goes straight to the backend.

This implements basic support for 32bit big endian bootloader;
tested on GRUB.

This prohibits writing to the block device as a precaution.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>